+++ SchoolManagerWPF/View/TabPages/Classes.xaml.cs
@@ -26,9 +26,9 @@ namespace SchoolManagerWPF.View.TabPages
             ClassesGrid.DataContext = getNewClassesViewModel(resourceManager);
         }
 
-        private ClassesViewModel getNewClassesViewModel(ResourceManager resourceManager)
+        private AdminClassesViewModel getNewClassesViewModel(ResourceManager resourceManager)
         {
-            var classesViewModel = new ClassesViewModel(resourceManager);
+            var classesViewModel = new AdminClassesViewModel(resourceManager);
             classesViewModel.SuccessfulClassAdd = () =>
             {
                 MessageBox.Show(resourceManager.GetString("SuccessfullyAdded"));
